Created variables for the coins
Doubles created to  get the values of gold coins silver coins and the left over which will be bronze coins.

The gold coins are calculated by dividing the totalAsDouble Variable by 10

The silver coins are calculate by dividing the leftOver variable by 5 and the bronze are found from the remainder using a modulo.